#673-circular-dependencies - fix compilation warnings
