+++ src/app/security-group/sg-list/security-group-list.component.ts
@@ -1,11 +1,18 @@
-import { Component, Input, OnChanges } from '@angular/core';
-import { SecurityGroup, SecurityGroupType } from '../sg.model';
+import {
+  Component,
+  Input,
+  OnChanges
+} from '@angular/core';
+import {
+  SecurityGroup,
+  SecurityGroupType
+} from '../sg.model';
 import { ViewMode } from '../../shared/components/view-mode-switch/view-mode-switch.component';
 import { SecurityGroupCardItemComponent } from '../sg-list-item/card-item/security-group-card-item.component';
 import { SecurityGroupRowItemComponent } from '../sg-list-item/row-item/security-group-row-item.component';
 import { TranslateService } from '@ngx-translate/core';
-import { SecurityGroupViewMode } from '../sg-filter/containers/sg-filter.container';
 import { ListService } from '../../shared/components/list/list.service';
+import { SecurityGroupViewMode } from '../sg-view-mode';
 
 
 @Component({
