+++ src/app/security-group/sg-view-mode.ts
@@ -1,0 +1,4 @@
+export enum SecurityGroupViewMode {
+  Templates = 'templates',
+  Shared = 'shared'
+}
