+++ src/app/shared/models/template-group.model.ts
@@ -5,3 +5,5 @@ export class TemplateGroup {
   constructor(id: string) {
   }
 }
+
+export const DefaultTemplateGroupId = 'general';
