+++ src/app/snapshot/recurring-snapshots/policy-list/policy-list.component.ts
@@ -9,11 +9,13 @@ import {
 } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
 import { TimeFormat } from '../../../shared/services/language.service';
-import { Policy, TimePolicy } from '../policy-editor/policy-editor.component';
-import { PolicyType } from '../recurring-snapshots.component';
+import {
+  Policy,
+  TimePolicy
+} from '../policy-editor/policy-editor.component';
 import { PolicyViewBuilderService } from './policy-view-builder.service';
+import { PolicyType } from '../snapshot-policy-type';
 import DateTimeFormat = Intl.DateTimeFormat;
-import { TableDatabase, TableDataSource } from '../../../shared/components/table/table';
 
 
 interface PolicyView {
