+++ src/app/snapshot/recurring-snapshots/snapshot-policy-type.ts
@@ -1,0 +1,6 @@
+export enum PolicyType {
+  Hourly,
+  Daily,
+  Weekly,
+  Monthly
+}
