+++ src/app/snapshot/recurring-snapshots/snapshot-policy.service.ts
@@ -4,11 +4,14 @@ import { BackendResource } from '../../shared/decorators';
 import { BaseBackendService } from '../../shared/services/base-backend.service';
 import { padStart } from '../../shared/utils/padStart';
 import { DayPeriod } from './day-period/day-period.component';
-import { Policy, TimePolicy } from './policy-editor/policy-editor.component';
-import { PolicyType } from './recurring-snapshots.component';
+import {
+  Policy,
+  TimePolicy
+} from './policy-editor/policy-editor.component';
 import { SnapshotPolicy } from './snapshot-policy.model';
 import { Time } from './time-picker/time-picker.component';
 import { HttpClient } from '@angular/common/http';
+import { PolicyType } from './snapshot-policy-type';
 
 
 export interface SnapshotPolicyCreationParams {
