+++ src/app/template/template-sidebar/template-group/template-group.component.ts
@@ -1,15 +1,17 @@
-import { Component, EventEmitter, Input, Output } from '@angular/core';
+import {
+  Component,
+  EventEmitter,
+  Input,
+  Output
+} from '@angular/core';
 import { MatDialog } from '@angular/material';
 import { BaseTemplateModel } from '../../shared/base-template.model';
-import { TemplateGroupSelectorComponent } from './template-group-selector/template-group-selector.component';
-import { TemplateGroupService } from '../../../shared/services/template-group.service';
 import { TranslateService } from '@ngx-translate/core';
 import { Language } from '../../../shared/services/language.service';
 import { TemplateTagKeys } from '../../../shared/services/tags/template-tag-keys';
 import { TemplateGroupSelectorContainerComponent } from './containers/template-group-selector.container';
 import { TemplateGroup } from '../../../shared/models/template-group.model';
 
-export const DefaultTemplateGroupId = 'general';
 
 @Component({
   selector: 'cs-template-group',
